+++ __tests__/rules/substituteFontAtRules.test.js
@@ -67,3 +67,48 @@ it('parses @font with size', () => {
     expect(result.warnings().length).toBe(0)
   })
 })
+
+it('parses @font with size and breakpoints', () => {
+  const input = `
+    article {
+      @font serif xs/1.53 >=lg;
+      @font serif sm(1.2)/1.53 <=sm;
+    }
+  `
+
+  const output = `
+    article {
+      font-family: Georgia,Cambria,"Times New Roman",Times,serif;
+      font-family: Georgia,Cambria,"Times New Roman",Times,serif;
+    }
+    @media (min-width: 1399px) and (max-width: 1899px) {
+      article {
+        font-size: 12px;
+        line-height: 1.53;
+      }
+    }
+    @media (min-width: 1900px) {
+      article {
+        font-size: 14px;
+        line-height: 1.53;
+      }
+    }
+    @media (min-width: 0) and (max-width: 739px) {
+      article {
+        font-size: 16.8px;
+        line-height: 1.53;
+      }
+    }
+    @media (min-width: 740px) and (max-width: 1023px) {
+      article {
+        font-size: 16.8px;
+        line-height: 1.53;
+      }
+    }
+  `
+
+  return run(input).then(result => {
+    expect(result.css).toMatchCSS(output)
+    expect(result.warnings().length).toBe(0)
+  })
+})