+++ __tests__/rules/substituteRowAtRules.test.js
@@ -99,14 +99,21 @@ it('parses @row with specified children count', () => {
   })
 })
 
-it('parses @row with query', () => {
+it('parses @row under @responsive', () => {
   const input = `
     article {
-      @row 3 xs;
+      @responsive xs {
+        @row 3;
+      }
     }
   `
 
   const output = `
+  @media (min-width: 0) and (max-width: 739px) {
+    article > * {
+      margin-left: 25px;
+    }
+  }
   @media (min-width: 0) and (max-width: 739px) {
     article {
       display: flex;
@@ -114,8 +121,81 @@ it('parses @row with query', () => {
     article > *:nth-child(3n+1) {
       margin-left: 0;
     }
-    article > * {
+  }
+  `
+
+  return run(input).then(result => {
+    expect(result.css).toMatchCSS(output)
+    expect(result.warnings().length).toBe(0)
+  })
+})
+
+it('parses @row as child under @responsive', () => {
+  const input = `
+    .inner {
+      article {
+        @responsive xs {
+          @row 3;
+        }
+      }
+    }
+  `
+
+  const output = `
+  @media (min-width: 0) and (max-width: 739px) {
+    .inner article > * {
       margin-left: 25px;
+    }
+  }
+  @media (min-width: 0) and (max-width: 739px) {
+    .inner article {
+      display: flex;
+    }
+    .inner article > *:nth-child(3n+1) {
+      margin-left: 0;
+    }
+  }
+  `
+
+  return run(input).then(result => {
+    expect(result.css).toMatchCSS(output)
+    expect(result.warnings().length).toBe(0)
+  })
+})
+
+it('parses @row as child under @responsive w/ advanced bpQuery', () => {
+  const input = `
+    .inner {
+      article {
+        @responsive $desktop {
+          @row 3;
+        }
+      }
+    }
+  `
+
+  const output = `
+  @media (min-width: 1024px) and (max-width: 1398px) {
+    .inner article > * {
+      margin-left: 50px;
+    }
+  }
+  @media (min-width: 1399px) and (max-width: 1899px) {
+    .inner article > * {
+      margin-left: 50px;
+    }
+  }
+  @media (min-width: 1900px) {
+    .inner article > * {
+      margin-left: 60px;
+    }
+  }
+  @media (min-width: 1024px) and (max-width: 1398px), (min-width: 1399px) and (max-width: 1899px), (min-width: 1900px) {
+    .inner article {
+      display: flex;
+    }
+    .inner article > *:nth-child(3n+1) {
+      margin-left: 0;
     }
   }
   `